Refactor(minjae/modal): Modal 리팩토링 #4
